+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-## 0.0.1
+# 1.0.0
 
-* TODO: Describe initial release.
+* Query: Search for volumes that contain this text string
+* PrintType: You can use the printType parameter to restrict the returned results to a specific print or publication type
+* Filter: You can use the filter parameter to restrict the returned results
+* Sorting: By default, a volume search request returns maxResults results, where maxResults is the parameter used in pagination (above), ordered by relevance to search terms.
+* Pagination: You can use the [StartIndex] and [MaxResults] parameters